Clean up all Python 3.6 warnings
Fixed a few Python3.6 deprecation warnings by replacing ``StopIteration``
with ``return``, as well as using ``getfullargspec()`` instead of
``getargspec()`` under Python 3.

Additionally fixed docstrings with backslashes needing r'', filehandles
not explicitly closed, accommodate for .pyc files not necessarily present.

Fixes: #458

+++ tests/test_script_consumption.py
@@ -536,7 +536,7 @@ def downgrade():
 
 """)
         pyc_path = util.pyc_file_from_path(path)
-        if os.access(pyc_path, os.F_OK):
+        if pyc_path is not None and os.access(pyc_path, os.F_OK):
             os.unlink(pyc_path)
 
         assert_raises_message(
